Move sequence_otel file to otel_to_pv
